bot/requirements.txt
python-telegram-bot==20.7
fastapi==0.105.0
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.2
requests==2.31.0 